+++ src/UxROM.cpp
@@ -9,7 +9,7 @@ inline uint32_t UxROM::CPU_READ(uint16_t ADDR) {
 }
 
 
-inline uint16_t UxROM::PPU_READ(uint16_t ADDR, bool NT) {
+inline uint32_t UxROM::PPU_READ(uint16_t ADDR, bool NT) {
 
     return (NT) ? SelectNameTable(ADDR, NT_MIRROR) : ADDR;
 }
